move create_queryset_by_queries
