+++ person/views.py
@@ -4,7 +4,8 @@ from typing import Callable, Iterable, Type, Any, NoReturn, TYPE_CHECKING, List
 
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin, PermissionRequiredMixin
-from django.db.models import F, Q
+from django.db import models
+from django.db.models import F, Q, OuterRef
 from django.db.models.lookups import LessThanOrEqual, GreaterThanOrEqual, Exact
 from django.forms import BaseForm
 from django.shortcuts import render, redirect, get_object_or_404
@@ -12,7 +13,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from cllib_django import query_utils
 from core import constant
 from core.constant import REL_TYPE_COMMENT_REFERS_TO, REL_TYPE_WAS_BORN_IN_LOCATION, REL_TYPE_DIED_AT_LOCATION, \
-    TRUE_CHAR
+    TRUE_CHAR, REL_TYPE_MENTION
 from core.export_data import cell_values, download_csv_serv
 from core.forms import CommentForm, PersonRecrefForm
 from core.helper import renderer_serv, view_serv, query_serv, recref_serv, form_serv, perm_serv
@@ -34,8 +35,9 @@ from person.models import CofkUnionPerson, CofkPersonPersonMap, create_person_id
 from person.person_serv import DisplayablePerson, SearchResultPerson
 from person.recref_adapter import PersonCommentRecrefAdapter, PersonResourceRecrefAdapter, PersonRoleRecrefAdapter, \
     ActivePersonRecrefAdapter, PassivePersonRecrefAdapter, PersonImageRecrefAdapter, PersonLocRecrefAdapter
-from person.subqueries import create_queryset_by_queries
+from person.subqueries import create_sql_count_work_by_person
 from person.view_components import PersonFormDescriptor
+from work.forms import AuthorRelationChoices, AddresseeRelationChoices
 
 if TYPE_CHECKING:
     from core.helper.view_serv import MergeChoiceContext
@@ -568,3 +570,44 @@ def lookup_other_details(lookup_fn, f, v) -> Q:
 
 def get_target_or_related_id(recref: CofkPersonPersonMap):
     return recref.related_id
+
+
+def create_queryset_by_queries(model_class: Type[models.Model], queries: Iterable[Q] = None, sort_by=None):
+    queryset = model_class.objects
+    annotate = {
+        'sent': create_sql_count_work_by_person(AuthorRelationChoices.values),
+        'recd': create_sql_count_work_by_person(AddresseeRelationChoices.values),
+        'all_works': create_sql_count_work_by_person(
+            AuthorRelationChoices.values + AddresseeRelationChoices.values),
+        'mentioned': create_sql_count_work_by_person([REL_TYPE_MENTION]),
+
+        'rolenames': CofkUnionPerson.objects.filter(
+            cofkpersonrolemap__person_id=OuterRef('pk'),
+        )
+        .annotate(_rolenames=query_utils.join_values_for_search('cofkpersonrolemap__role__role_category_desc'))
+        .values_list('_rolenames', flat=True),
+        'names_and_titles': query_utils.concat_safe(
+            [
+                'foaf_name',
+                'skos_altlabel',
+                'person_aliases',
+                'rolenames',
+            ]
+        )
+
+    }
+
+    queryset = query_serv.update_queryset(queryset, model_class, queries=queries,
+                                          annotate=annotate, sort_by=sort_by)
+    queryset = queryset.prefetch_related(
+        'cofkpersonlocationmap_set__location',
+        'cofkpersoncommentmap_set__comment',
+        'roles',
+        'resources',
+        'images',
+        'comments',
+        'active_relationships__related',
+        'passive_relationships__person',
+    )
+
+    return queryset
